projects/erp_ganadero/backend/requirements.txt
# ERP Ganadero - Backend Requirements

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
supabase==2.3.0
structlog==24.1.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.24.1

# Type checking
mypy==1.8.0
